lines

+++ linesStops/models.py
@@ -6,7 +6,7 @@ class Lines (models.Model) :
     line = models.CharField(max_length=50)
     
 class Stations (models.Model) : 
-    st_latitude=models.IntegerField()
-    st_longitude=models.IntegerField ()
+    st_latitude=models.FloatField()
+    st_longitude=models.FloatField ()
     station = models.CharField(max_length=50)
     line =  models.ForeignKey(Lines, on_delete=models.CASCADE)
